Make Bake find itself.
It's common for each build project to have it's own copy of Bake,
requiring one to cd to the correct place to invoke any of the Bake
commands.

This patch enables Bake to find itself.

This enables you to put one copy of Bake on your PATH, then simply
invoke that  The patch will cause Bake to walk up the directory tree
from the current working directory until it finds the root of the
project, then invoke the Bake found there.  This is similar to how
git can be invoked anywhere in the project tree and still finds the
.git directory at the root.